modified:   movie/src/main/java/com/example/movie/controller/MovieController.java 	modified:   movie/src/main/java/com/example/movie/model/Movie.java 	modified:   movie/src/main/java/com/example/movie/repository/MovieRepository.java 	modified:   movie/src/main/java/com/example/movie/service/MovieService.java 	modified:   movie/src/main/java/com/example/movie/service/MovieServiceImpl.java 	modified:   movie/src/main/resources/static/app.js 	modified:   movie/src/main/resources/static/index.html 	modified:   movie/src/main/resources/static/styles.css

+++ movie/src/main/java/com/example/movie/service/MovieService.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface MovieService {
     List<Movie> getAllMovies();
-    Movie getMovieById(int id);
+    Movie getMovieById(Long id);
     Movie saveMovie(Movie movie);
-    void deleteMovie(int id);
+    void deleteMovie(Long id);
 }
